Refactored JoinServer logic
